fix(dns): redact permission diagnostics
